Isolate build from host environment variables
The build should not be influenced by things such as `LANG` set in the
host; so we disable inheritance of `ENV` by manually calling `setenv()`
every time.  This unfortunately alters the signature of both the `do`
block when calling `debootstrap()` and `chroot()`, but luckily Julia's
dynamism makes it easy to work around.