Add `zstd` to `agent_linux` images (#238)
While the outer agent images are intentionally quite bare, the compression/decompression speed of `zstd` makes it extremely attractive to use by default within plugins.  I intend to use it in `coppermind` by default, so I'm adding it to the bare agent image for now, with a future "JLL installation plugin" to come.